add print html page functionality
