+++ apps/web/src/components/ui/report/hooks/buildExportHtml2.ts
@@ -1,5 +1,9 @@
+import { PAGE_CONTROLLER_ID } from '@/controllers/ReportPageControllers/ReportPageController';
 import type { PlateEditor } from 'platejs/react';
+import { ReportPageController } from '@/controllers/ReportPageControllers/ReportPageController';
 
-export const buildExportHtml = async (editor: PlateEditor, options?: {}): Promise<string> => {
-  return '';
+export const useBuildExportHtml2 = () => {
+  return async ({ reportId }: { reportId: string }) => {
+    return '';
+  };
 };
